ref: Refactored code according to comments
