+++ src/components/MoreDetailsBtn.tsx
@@ -4,7 +4,6 @@ import add from '../../../assets/icons/nav/arrow-right.svg'
 
 export default function MoreDetailsBtn() {
     return (
-        // Still need to include an destination link
         <button className='flex pl-4 pr-2 gap-2 h-[40px] items-center border-solid border-white border rounded-lg font-XS-bold'>
             Ver mais
             <Image
