+++ src/types/modules/app/components/FAQItem.d.ts
@@ -1,0 +1,4 @@
+export type FAQItemProps = {
+    question: string,
+    answer: string
+};
