+++ src/utils/faqQuestions.ts
@@ -1,0 +1,6 @@
+export const questions = [
+    { question: 'O que é TableRise?', answer: 'É uma plataforma VTT (Virtual Table Top) para jogar RPG online com amigos.' },
+    { question: 'TableRise é pago?', answer: 'Nâo, todas nossas funcionalidades são gratuitas.' },
+    { question: 'Quais modos de jogo existem?', answer: 'Em um primeiro momento é possível jogar apenas D&D e sem precisar adquirir um livro, porém já está planejado desenvolvermos um modo mais geral, que se encaixa com qualquer sistema de RPG.' },
+    { question: 'O conteúdo do D&D é original da Wizard?', answer: 'Por sermos um projeto sem fins lucrativos utilizamos o conteúdo disponibilizado pela Wizard para utilização gratuita (CC), porém o conteúdo do SRD da Wizard não é suficiente para uma campanha real, por isso em alguns pontos nos do TableRise tomamos a liberdade de adicionarmos alguns conteúdos autorais. A proporção é de 70% WIzard para 30% TableRise.' },
+];
